MTOCISA

+++ src/br/ufsc/ine5605/trabalho1/controle/ControladorEleitor.java
@@ -3,6 +3,7 @@ package br.ufsc.ine5605.trabalho1.controle;
 import br.ufsc.ine5605.trabalho1.entidade.Eleitor;
 import br.ufsc.ine5605.trabalho1.apresentacao.TelaEleitorOLD;
 import br.ufsc.ine5605.trabalho1.apresentacao.TelaEleitor;
+import br.ufsc.ine5605.trabalho1.exception.EleitorDuplicado;
 import br.ufsc.ine5605.trabalho1.mapeador.Mapeador;
 import java.util.ArrayList;
 
@@ -13,7 +14,15 @@ public class ControladorEleitor implements IControlador<Eleitor> {
 
     private ControladorEleitor() {
         this.mapper = new Mapeador<>("eleitores.urn");
+        load();
+    }
+
+    private void load() {
         mapper.load();
+
+        for (Eleitor x : mapper.getList()) {
+            x.setCidade(ControladorCidade.getInstance().getCidade(x.getCidade().getCodigo()));
+        }
     }
 
     public void persist() {
@@ -29,10 +38,10 @@ public class ControladorEleitor implements IControlador<Eleitor> {
     }
 
     @Override
-    public boolean cadastra(Eleitor eleitor) {
+    public boolean cadastra(Eleitor eleitor) throws EleitorDuplicado {
         for (Eleitor eleitorCadastrado : mapper.getList()) {
             if (eleitorCadastrado.getTitulo() == eleitor.getTitulo()) {
-                return false;
+                throw new EleitorDuplicado();
             }
         }
 
@@ -45,7 +54,10 @@ public class ControladorEleitor implements IControlador<Eleitor> {
     }
 
     @Override
-    public boolean modifica(Eleitor antigo, Eleitor novo) {
+    public boolean modifica(Eleitor antigo, Eleitor novo) throws EleitorDuplicado {
+        if (getEleitor(novo.getTitulo()) != null)
+            throw new EleitorDuplicado();
+        
         if (mapper.contains(antigo.getTitulo())) {
 
             mapper.remove(antigo.getTitulo());
@@ -54,7 +66,7 @@ public class ControladorEleitor implements IControlador<Eleitor> {
             antigo.setTitulo(novo.getTitulo());
             antigo.setSecaoEleitoral(novo.getSecaoEleitoral());
             antigo.setZonaEleitoral(novo.getZonaEleitoral());
-          
+
             mapper.put(antigo.getTitulo(), antigo);
             return true;
         }
@@ -67,13 +79,7 @@ public class ControladorEleitor implements IControlador<Eleitor> {
     }
 
     public Eleitor getEleitor(long titulo) {
-        for (Eleitor eleitor : mapper.getList()) {
-            if (eleitor.getTitulo() == titulo) {
-                return eleitor;
-            }
-        }
-
-        return null;
+        return mapper.get(titulo);
     }
 
     public Eleitor getEleitor(String nome) {
